refactor: 컴포넌트 및 유틸 분리
