+++ src/app/(home)/@bot/components/PromotionReviews.tsx
@@ -5,6 +5,8 @@ import { DEFAULT_SSG_REVALIDATE_TIME } from '@/constants/common';
 import Article from '@/components/article/Article';
 import { ReviewsViewEntity } from '@/types/review.type';
 import UserProfile from '@/components/header/UserProfile';
+import { FeedbackGroup } from '@/components/review/FeedbackGroup';
+import { MonthlyReviewChip } from '@/components/review/MonthlyReviewChip';
 
 export const revalidate = DEFAULT_SSG_REVALIDATE_TIME;
 
@@ -57,58 +59,3 @@ const PromotionReview = async () => {
 };
 
 export default PromotionReview;
-
-export const MonthlyReviewChip = () => {
-  return (
-    <div className="rounded-[42px] bg-basic-grey-50 px-8 py-4 text-10 font-600 leading-[160%] text-basic-grey-700">
-      이달의 후기
-    </div>
-  );
-};
-
-export const FeedbackGroup = () => {
-  return (
-    <div className="flex gap-[6px] ">
-      <PassengerRegion />
-      <p className="text-12 font-500 leading-[160%] text-basic-grey-200">|</p>
-      <FeedbackCard type="서비스" text="매우 만족" />
-      <p className="text-12 font-500 leading-[160%] text-basic-grey-200">|</p>
-      <FeedbackCard type="탑승" text="매우 만족" />
-    </div>
-  );
-};
-
-export const PassengerRegion = () => {
-  return (
-    <p className="text-12 font-500 leading-[160%] text-basic-grey-700">
-      경남 탑승객
-    </p>
-  );
-};
-
-export const FeedbackCard = ({
-  type,
-  text,
-}: {
-  type: FeedbackType;
-  text: FeedbackText;
-}) => {
-  return (
-    <div className="flex items-center gap-4">
-      <p className="text-12 font-500 leading-[160%] text-basic-grey-700">
-        {type}
-      </p>
-      <p className="text-12 font-500 leading-[160%] text-basic-grey-500">
-        {text}
-      </p>
-    </div>
-  );
-};
-
-export type FeedbackType = '서비스' | '탑승';
-export type FeedbackText =
-  | '매우 불만족'
-  | '불만족'
-  | '보통'
-  | '만족'
-  | '매우 만족';
