+++ src/components/review/MonthlyReviewChip.tsx
@@ -1,0 +1,7 @@
+export const MonthlyReviewChip = () => {
+  return (
+    <div className="rounded-[42px] bg-basic-grey-50 px-8 py-4 text-10 font-600 leading-[160%] text-basic-grey-700">
+      이달의 후기
+    </div>
+  );
+};
